Create a FrozenClock that always returns the same date in the past
This was taken from my usaidwat project. FrozenClock::default() returns
a clock with a notion of now that is always the same time: 23 May 2025
at 10:13 AM PDT. That was the same time it was defined to return in
usaidwat, so horologe will be a drop-in replacement for the equivalent
code in usaidwat.

I'd like to make a macro that can implement Default with a given
datetime defined by each user of this crate, but that's a task for
later.